Refactor the HostedClusterReconciler to be more organized
Add "platform reconcilers" to implement the new `PlatformReconciler` interface, for several reasons:
1. the platform specific code is now in a dedicated files.
2. to ease of adding new platforms support
3. to cancel the switch-cases in the code, to simplify it and make it more maintainable.
