+++ hypershift-operator/controllers/hostedcluster/hostedcluster_controller_test.go
@@ -4,12 +4,10 @@ import (
 	"testing"
 	"time"
 
-	"github.com/openshift/hypershift/hypershift-operator/controllers/manifests/ignitionserver"
-	"k8s.io/apimachinery/pkg/util/intstr"
-	capiawsv1 "sigs.k8s.io/cluster-api-provider-aws/api/v1alpha4"
-
 	"github.com/google/go-cmp/cmp"
 	. "github.com/onsi/gomega"
+	"github.com/openshift/hypershift/hypershift-operator/controllers/manifests/ignitionserver"
+	"k8s.io/apimachinery/pkg/util/intstr"
 
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/equality"
@@ -713,91 +711,6 @@ func TestReconcileIgnitionServerServiceRoute(t *testing.T) {
 			g.Expect(test.inputIgnitionServerService.Spec.Ports[0].Port).To(Equal(int32(443)))
 			g.Expect(test.inputIgnitionServerService.Spec.Ports[0].Name).To(Equal("https"))
 			g.Expect(test.inputIgnitionServerService.Spec.Ports[0].Protocol).To(Equal(corev1.ProtocolTCP))
-		})
-	}
-}
-
-func TestReconcileAWSCluster(t *testing.T) {
-	testCases := []struct {
-		name              string
-		initialAWSCluster *capiawsv1.AWSCluster
-		hostedCluster     *hyperv1.HostedCluster
-
-		expectedAWSCluster *capiawsv1.AWSCluster
-	}{
-		{
-			name:              "Tags get copied over",
-			initialAWSCluster: &capiawsv1.AWSCluster{},
-			hostedCluster: &hyperv1.HostedCluster{Spec: hyperv1.HostedClusterSpec{Platform: hyperv1.PlatformSpec{AWS: &hyperv1.AWSPlatformSpec{
-				ResourceTags: []hyperv1.AWSResourceTag{
-					{Key: "foo", Value: "bar"},
-				},
-			}}}},
-
-			expectedAWSCluster: &capiawsv1.AWSCluster{
-				ObjectMeta: metav1.ObjectMeta{Annotations: map[string]string{
-					"cluster.x-k8s.io/managed-by":     "external",
-					"hypershift.openshift.io/cluster": "/",
-				}},
-				Spec: capiawsv1.AWSClusterSpec{
-					AdditionalTags: capiawsv1.Tags{"foo": "bar"},
-				},
-				Status: capiawsv1.AWSClusterStatus{
-					Ready: true,
-				},
-			},
-		},
-		{
-			name: "Existing tags get removed",
-			initialAWSCluster: &capiawsv1.AWSCluster{Spec: capiawsv1.AWSClusterSpec{AdditionalTags: capiawsv1.Tags{
-				"to-be-removed": "value",
-			}}},
-			hostedCluster: &hyperv1.HostedCluster{Spec: hyperv1.HostedClusterSpec{Platform: hyperv1.PlatformSpec{AWS: &hyperv1.AWSPlatformSpec{
-				ResourceTags: []hyperv1.AWSResourceTag{
-					{Key: "foo", Value: "bar"},
-				},
-			}}}},
-
-			expectedAWSCluster: &capiawsv1.AWSCluster{
-				ObjectMeta: metav1.ObjectMeta{Annotations: map[string]string{
-					"cluster.x-k8s.io/managed-by":     "external",
-					"hypershift.openshift.io/cluster": "/",
-				}},
-				Spec: capiawsv1.AWSClusterSpec{
-					AdditionalTags: capiawsv1.Tags{"foo": "bar"},
-				},
-				Status: capiawsv1.AWSClusterStatus{
-					Ready: true,
-				},
-			},
-		},
-		{
-			name: "No tags on hostedcluster clears existing awscluster tags",
-			initialAWSCluster: &capiawsv1.AWSCluster{Spec: capiawsv1.AWSClusterSpec{AdditionalTags: capiawsv1.Tags{
-				"to-be-removed": "value",
-			}}},
-			hostedCluster: &hyperv1.HostedCluster{},
-
-			expectedAWSCluster: &capiawsv1.AWSCluster{
-				ObjectMeta: metav1.ObjectMeta{Annotations: map[string]string{
-					"cluster.x-k8s.io/managed-by":     "external",
-					"hypershift.openshift.io/cluster": "/",
-				}},
-				Status: capiawsv1.AWSClusterStatus{
-					Ready: true,
-				},
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			if err := reconcileAWSCluster(tc.initialAWSCluster, tc.hostedCluster, hyperv1.APIEndpoint{}); err != nil {
-				t.Fatalf("reconcileAWSCluster failed: %v", err)
-			}
-			if diff := cmp.Diff(tc.initialAWSCluster, tc.expectedAWSCluster); diff != "" {
-				t.Errorf("reconciled AWS cluster differs from expcted AWS cluster: %s", diff)
-			}
 		})
 	}
 }
